selftest: acct: Add selftest for the acct() syscall
Noticed that there was no selftest for the acct() syscall
which enables the kernel to record terminated processes
into a specified file.

The acct() system call enables or disables process accounting.
If accounting is turned on, records for each terminating process
are appended to a specified filename as it terminates. An argument of NULL
causes accounting to be turned off.

This patch provides a test for the acct() syscall.

References:
https://man7.org/linux/man-pages/man2/acct.2.html

Signed-off-by: Abdulrasaq Lawani <abdulrasaqolawani@gmail.com>